+++ BGME.Framework/P4G/Sound.cs
@@ -39,7 +39,7 @@ internal unsafe class Sound : BaseSound
     {
         scanner.Scan("Play Sound Function", "48 63 E9 89 D6", result =>
         {
-            this.playSoundFunction = hooks.CreateFunction<PlaySoundFunction>((nint)result - 30);
+            this.playSoundFunction = hooks.CreateFunction<PlaySoundFunction>(result - 30);
             this.playSoundHook = this.playSoundFunction.Hook(this.PlaySoundImpl).Activate();
         });
